♻️ Github DB의 context이름을 content이름으로 수정하며 내부 로직의 CodingKeys 제거
