+++ Workade/Models/CheckListTemplateResource.swift
@@ -1,5 +1,5 @@
 //
-//  CheckListTemplate.swift
+//  CheckListTemplateResource.swift
 //  Workade
 //
 //  Created by 김예훈 on 2022/10/26.
@@ -12,10 +12,6 @@ struct CheckListTemplateResource: Codable {
     
     init() {
         self.content = []
-    }
-    
-    enum CodingKeys: String, CodingKey {
-        case content = "context"
     }
 }
