+++ Workade/Models/MagazineDetailResource.swift
@@ -1,5 +1,5 @@
 //
-//  MagazineModel.swift
+//  MagazineDetailResource.swift
 //  Workade
 //
 //  Created by Hong jeongmin on 2022/10/27.
@@ -22,9 +22,4 @@ struct MagazineDetailModel: Codable {
     let font: String?
     let content: String
     let color: String?
-    
-    enum CodingKeys: String, CodingKey {
-        case type, font, color
-        case content = "context"
-    }
 }
